LP-2413 Moving utils from common to lms
